# openapi.yaml
openapi: 3.1.0
info:
  title: YouTube Video Metadata API
  description: A REST API to extract YouTube video metadata
  version: 0.1.0
servers:
  - url: '{BASE_URL}'
    description: Configurable server (set BASE_URL environment variable)
    variables:
      BASE_URL:
        default: http://localhost:8000
        description: Base URL for the API server (configure via BASE_URL environment variable)
paths:
  /api/video-metadata:
    get:
      summary: Get Youtube Video Info Get
      description: Get YouTube video metadata such as title, description, and thumbnail using the full video URL
      operationId: get_youtube_video_metadata
      parameters:
        - name: url
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            description: YouTube video URL
            title: Url
          description: YouTube video URL
          example: https://www.youtube.com/watch?v=ZHYK0m9rpB0
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoMeta'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
    VideoMeta:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        thumbnail_url:
          type: string
          title: Thumbnail Url
        channel:
          type: string
          title: Channel
        channel_url:
          type: string
          title: Channel Url
        duration:
          type: integer
          title: Duration
        view_count:
          type: integer
          title: View Count
      type: object
      required:
        - id
        - title
        - description
        - thumbnail_url
        - channel
        - channel_url
        - duration
        - view_count
      title: VideoMeta